refactor: 404 페이지 수정
- ErrorImageGroup → ErrorIllustration
- Error404Page → NotFoundPage
